refactor: share parseText helper
