+++ src/NodeCard.jsx
@@ -1,23 +1,6 @@
 import { memo } from 'react'
 import { Handle, Position } from 'reactflow'
-import { marked } from 'marked'
-
-function parseText(text = '') {
-  const tokens = marked.lexer(text)
-  let title = ''
-  let foundTitle = false
-  const bodyParts = []
-  for (const t of tokens) {
-    if (!foundTitle && t.type === 'heading') {
-      title = t.text
-      foundTitle = true
-    } else if (t.type === 'paragraph' || t.type === 'text') {
-      bodyParts.push(t.text)
-    }
-  }
-  const body = bodyParts.join(' ').trim()
-  return { title, snippet: body.slice(0, 50) }
-}
+import { parseText } from './parseText.js'
 
 const NodeCard = memo(({ id, data, selected }) => {
   const { title, snippet } = parseText(data.text)
